requirements.txt
aioredis==2.0.1
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.26.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.3
cryptography==38.0.1
databases==0.6.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
environs==9.5.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-pagination==0.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
Jinja2==3.1.4
Mako==1.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mongoengine==0.28.2
motor==3.4.0
orjson==3.10.3
packaging==24.0
passlib==1.7.4
psycopg2-binary==2.9.4
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.2
Pygments==2.18.0
PyJWT==2.6.0
pymongo==4.7.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.3.4
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.37.2
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.12.1
ujson==5.10.0
uvicorn==0.18.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
beanie==1.26.0
